requirements.txt
discord.py==2.7.0
yt-dlp==2023.12.06
sqlite3
